+++ trunk/src/core/srs_core_deprecated.cpp
@@ -1,0 +1,8 @@
+//
+// Copyright (c) 2013-2024 The SRS Authors
+//
+// SPDX-License-Identifier: MIT
+//
+
+#include <srs_core_deprecated.hpp>
+
